Move machine to common.
