+++ src/ts/common/machine.ts
